requirements.txt
ollama
pymupdf
python-docx
